requirements.txt
gspread
oauth2client
openai
python-dotenv
flask
gunicorn